move: base dependencies
