+++ app/entities/offer/routers_offer.py
@@ -14,10 +14,7 @@ from entities.offer.schemas_offer import (
     OfferLinkOptionalSchema,
 )
 from entities.offer.services_offer import OfferLinkService
-from dependencies.base_dependency import (
-    ValidGetByIdDependency,
-    valid_empty_schema,
-)
+from dependencies import valid_empty_schema, ValidGetByIdDependency
 
 router_offer = APIRouter()
 R_OFFER = '/offer/'
