+++ app/entities/users/routers/client_routers.py
@@ -6,10 +6,7 @@ from routers.consts import RouteSlug
 from models.database import get_session
 from entities.users.schemas_users import ClientInSchema, ClientInOptionalSchema
 from entities.users.services.client_service import ClientService
-from dependencies.base_dependency import (
-    ValidGetByIdDependency,
-    valid_empty_schema,
-)
+from dependencies import valid_empty_schema, ValidGetByIdDependency
 
 router_client = APIRouter()
 ROUTE_CLIENT = '/client/'
